import fix

+++ app/backend/helpers/Config.py
@@ -1,4 +1,5 @@
 import io
+import os
 import json
 
 
@@ -6,11 +7,17 @@ class Config:
     def __init__(self, path):
         self.path = path
         self.file = self._open_or_create_file(path)
+
         self.params = self._config_to_dict(self.file)
 
     def set_config(self, k, v):
         self.params[k] = v
         self.save_file()
+
+    def exists(self):
+        if os.path.exists(self.path):
+            return True
+        return False
 
     def get_config(self, k):
         try:
@@ -35,7 +42,10 @@ class Config:
         self.params = self._config_to_dict(self.file)
 
     def _open_or_create_file(self, path):
-        f = open(f"{path}", "w+")
+        if self.exists():
+            f = open(f"{path}", "r")
+        else:
+            f = open(f"{path}", "w+")
         return f
 
     def _config_to_dict(self, s):
